Put groupvel and led function into header file Improve readability and modularity
